requirements.txt
flask
prometheus-client
PyGithub
requests
